refactor: styles 구조화 리팩터링, media query를 위한 respondTo util 생성
