+++ src/components/Board/Board.jsx
@@ -1,6 +1,6 @@
 import { Link, useNavigate } from 'react-router-dom';
-import styled from 'styled-components';
-import { ellipsisStyle } from '@/styles/utils';
+import styled, { css } from 'styled-components';
+import { ellipsisStyle, respondTo } from '@/styles/utils';
 import { useInfiniteGetFeedbacks } from '@/stores/queries/useFeedbackQueries';
 import InfiniteScroll from '../InfiniteScroll';
 
@@ -46,9 +46,9 @@ const StContainerBox = styled.ul`
   gap: 24px;
   padding: 48px 0 32px;
 
-  @media (max-width: 1400px) {
+  ${respondTo.desktop(css`
     padding: 48px 20px;
-  }
+  `)}
 `;
 const StButton = styled.button`
   width: 60px;
