+++ src/components/Layout/Layout.jsx
@@ -1,6 +1,7 @@
 import { Outlet } from 'react-router-dom';
-import styled from 'styled-components';
+import styled, { css } from 'styled-components';
 import useScrollToTop from '@/hooks/useScrollToTop';
+import { respondTo } from '@/styles/theme';
 import Header from './Header';
 
 const Layout = () => {
@@ -21,9 +22,9 @@ const StLayout = styled.main`
   margin: 0 auto;
   width: 100%;
 
-  @media (max-width: 1400px) {
+  ${respondTo.desktop(css`
     padding: 0 20px;
-  }
+  `)}
 `;
 
 export default Layout;
